pyproject.toml
[tool.poetry]
name = "minecraft-whitelister"
version = "0.0.0"
description = "A Discord bot to whitelist members on a Java Edition server (with Geyser)"
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.20.0"
discord-py = "^2.2.3"
arrow = "^1.2.3"
aio-mc-rcon = "^3.2.0"
aiohttp = "^3.8.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.269"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
